Added comment features

+++ controllers/boardController.js
@@ -1,5 +1,5 @@
 const { v4: uuidv4 } = require("uuid");
-const { Board, User } = require("../models");
+const { Board, User, Task, BoardMember } = require("../models");
 
 exports.createBoard = async (req, res) => {
   const orgId = req.orgId;
@@ -12,7 +12,6 @@ exports.createBoard = async (req, res) => {
       description,
       orgId,
       stageOrder: [],
-      assignedMembers: userType === "org" ? [] : [req.user.user.id],
     });
 
     if (userType === "user") {
@@ -21,14 +20,22 @@ exports.createBoard = async (req, res) => {
         return res
           .status(404)
           .json({ message: "User not found", success: false });
-      await User.update(
-        {
-          associatedBoards: user.associatedBoards
-            ? [...user.associatedBoards, board.id]
-            : [board.id],
-        },
-        { where: { id: req.user.user.id } }
-      );
+
+      const boardMember = await BoardMember.findOne({
+        where: { userId: user.id, boardId: board.id },
+      });
+
+      if (boardMember)
+        return res.status(400).json({
+          message: "User already added",
+          success: false,
+        });
+
+      await BoardMember.create({
+        id: uuidv4(),
+        userId: user.id,
+        boardId: board.id,
+      });
     }
 
     res.status(201).json({
@@ -50,7 +57,10 @@ exports.updateBoard = async (req, res) => {
   const { title, description } = req.body;
   const orgId = req.orgId;
   try {
-    await Board.update({ title, description }, { where: { id: boardId } });
+    await Board.update(
+      { title, description },
+      { where: { id: boardId, orgId } }
+    );
 
     res.status(200).json({
       message: "Board updated successfully",
@@ -69,8 +79,9 @@ exports.updateBoard = async (req, res) => {
 exports.deleteBoard = async (req, res) => {
   const boardId = req.params.boardId;
   const orgId = req.orgId;
+  const board = req.board;
   try {
-    await Board.destroy({ where: { id: boardId } });
+    await Board.destroy({ where: { id: boardId, orgId } });
     res.status(200).json({
       message: "Board deleted successfully",
       success: true,
@@ -105,7 +116,6 @@ exports.getBoard = async (req, res) => {
 
 exports.getAllBoards = async (req, res) => {
   const orgId = req.orgId;
-  console.log("Logging from line 119");
   try {
     const allBoard = await Board.findAll(
       { attributes: { exclude: ["createdAt", "updatedAt"] } },
@@ -115,6 +125,13 @@ exports.getAllBoards = async (req, res) => {
       return res
         .status(404)
         .json({ message: "Boards not found", success: false });
+
+    if (allBoard.length === 0) {
+      return res.status(200).json({
+        message: "No boards found",
+        success: false,
+      });
+    }
 
     res.status(200).json({
       message: "Boards listed successfully",
@@ -134,13 +151,9 @@ exports.addMember = async (req, res) => {
   const { email } = req.body;
   const orgId = req.orgId;
   const boardId = req.params.boardId;
-  const userType = req.user.userType;
-  if (userType !== "org")
-    return res.status(401).json({ message: "Unauthorized", success: false });
   try {
     const board = await Board.findOne({ where: { id: boardId, orgId } });
     const user = await User.findOne({ where: { email, orgId } });
-
     if (!board)
       return res
         .status(404)
@@ -149,42 +162,19 @@ exports.addMember = async (req, res) => {
       return res
         .status(404)
         .json({ message: "User not found", success: false });
-
-    if (board.assignedMembers.includes(user.id)) {
+    const boardMemberData = await BoardMember.findOne({
+      where: { userId: user.id, boardId },
+    });
+    if (boardMemberData)
       return res.status(400).json({
         message: "User already added",
         success: false,
       });
-    }
-    await User.update(
-      {
-        associatedBoards: user.associatedBoards
-          ? [...user.associatedBoards, board.id]
-          : [board.id],
-      },
-      {
-        where: { id: user.id },
-      }
-    );
-    let updatedBoard;
-    if (board.assignedMembers.length === 0) {
-      updatedBoard = await Board.update(
-        { assignedMembers: [user.id] },
-        { where: { id: boardId, orgId } }
-      );
-    } else {
-      updatedBoard = await Board.update(
-        {
-          assignedMembers: [...board.assignedMembers, user.id],
-        },
-        {
-          where: {
-            id: boardId,
-            orgId,
-          },
-        }
-      );
-    }
+    await BoardMember.create({
+      id: uuidv4(),
+      userId: user.id,
+      boardId,
+    });
     res.status(200).json({
       message: "Member added successfully",
       success: true,
@@ -201,40 +191,28 @@ exports.addMember = async (req, res) => {
 exports.removeMember = async (req, res) => {
   const { userId, boardId } = req.params;
   const orgId = req.orgId;
-  const userType = req.user.userType;
-  if (userType !== "org")
-    return res.status(401).json({ message: "Unauthorized", success: false });
   try {
     const board = await Board.findOne({ where: { id: boardId, orgId } });
     if (!board)
       return res
         .status(404)
         .json({ message: "Board not found", success: false });
-    if (!board.assignedMembers.includes(userId)) {
+    const user = await User.findOne({ where: { id: userId, orgId } });
+    if (!user)
+      return res
+        .status(404)
+        .json({ message: "User not found", success: false });
+    const boardMemberData = await BoardMember.findOne({
+      where: { userId: user.id, boardId },
+    });
+    if (!boardMemberData)
       return res.status(400).json({
         message: "User not a board member",
         success: false,
       });
-    }
-    const user = await User.findOne({ where: { id: userId, orgId } });
-    await User.update(
-      {
-        associatedBoards: user.associatedBoards.filter(
-          (board) => board !== boardId
-        ),
-      },
-      {
-        where: { id: user.id },
-      }
-    );
-    const updatedBoard = await Board.update(
-      {
-        assignedMembers: board.assignedMembers.filter(
-          (member) => member !== userId
-        ),
-      },
-      { where: { id: boardId, orgId } }
-    );
+    await BoardMember.destroy({
+      where: { userId: user.id, boardId },
+    });
 
     res.status(200).json({
       message: "Member removed successfully",
@@ -242,6 +220,32 @@ exports.removeMember = async (req, res) => {
     });
   } catch (error) {
     console.log(error);
+    res.status(500).json({
+      message: "Something went wrong",
+      success: false,
+      error: error.message,
+    });
+  }
+};
+
+exports.orgAllBoards = async (req, res) => {
+  const orgId = req.orgId;
+  try {
+    const allBoard = await Board.findAll(
+      { attributes: { exclude: ["createdAt", "updatedAt"] } },
+      { where: { orgId } }
+    );
+    if (!allBoard)
+      return res
+        .status(404)
+        .json({ message: "Boards not found", success: false });
+
+    res.status(200).json({
+      message: "Boards listed successfully",
+      success: true,
+      data: allBoard,
+    });
+  } catch (error) {
     res.status(500).json({
       message: "Something went wrong",
       success: false,
